tests/wpt/tests/css/css-masonry/tentative/parsing/masonry-flow-computed.html

<!DOCTYPE html>
<html>
<head>
  <title>CSS Masonry: masonry-flow getComputedStyle()</title>
  <link rel="author" title="Celeste Pan" href="mailto:celestepan@microsoft.com">
  <link rel="help" href="https://drafts.csswg.org/css-grid-3">
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="/css/support/computed-testcommon.js"></script>
  <script src="/css/support/inheritance-testcommon.js"></script>
</head>
<body>
  <div id="target"></div>
  <script>
    test_computed_value("masonry-flow", "column normal");
    test_computed_value("masonry-flow", "column reverse");

    test_computed_value("masonry-flow", "row normal");
    test_computed_value("masonry-flow", "row reverse");

    test_computed_value("masonry-flow", "column-reverse normal");
    test_computed_value("masonry-flow", "column-reverse reverse");

    test_computed_value("masonry-flow", "row-reverse normal");
    test_computed_value("masonry-flow", "row-reverse reverse");
  </script>
</body>
</html>
